move transit available serializer to tyr
